aplicación ordenada
